requirements.txt
# Core dependencies
python-telegram-bot[job-queue]>=20.7
requests>=2.31.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
responses>=0.24.0

# TTS
google-cloud-texttospeech>=2.14.0
pydub>=0.25.1

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
types-requests>=2.31.0
